Add NPE protection to RemapperAdapter
In 59db546c227bce4abaf75be1d89bf06067966e98, IRemapper's contract was changed
so that mapMethodName and mapFieldName may be called with null arguments,
but RemapperAdapterFML was not changed to be able to handle this, causing a
NPE to happen in this case.

This commit adds an opt-in way for RemapperAdapter implementations to request
that the remapper should immediately no-op when invoked with null arguments,
and makes RemapperAdapterFML use it.