Don't add classpath mods to FML reparseable coremod collection
As proposed in https://github.com/SpongePowered/Mixin/issues/207

Fixes mods on the class path containing mixins not getting loaded as mods in a
1.7.10 development environment.

Background information:

Discovery of Forge mods roughly happens in the following 2 steps:

1. `findClassPathMods`: Discover Forge mods on the class path which are not in
the coremod list or the reparseable coremod list.

2. `findModDirMods`: Discover Forge mods that are in the mod directory, or
specified as *extra mods* using `--mods`.

It appears that the purpose of the reparseable mod list is to prevent a mod from
being discovered in both of the steps. This would happen with extra mods
containing both a coremod and a Forge mod, since they get injected into the
class path during coremod discovery (but since they get marked as reparseable,
this is avoided.)

Mixin needs to make sure the mod gets discovered in exactly one of the steps. We
can achieve this by only adding it to the reparseable list if it's not on the
boot class path. This causes only extra mods to get added.